+++ src/main/java/com/labs/tenderservice/service/User.java
@@ -1,15 +1,15 @@
 package com.labs.tenderservice.service;
 
 public class User {
-    private final int id;
+    private final long id;
     private final String username;
 
-    public User(int id, String username) {
+    public User(long id, String username) {
         this.id = id;
         this.username = username;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
